requirements.txt
Flask==1.1.1
beautifulsoup4==4.7.1
gevent==1.4.0
python-dotenv==0.10.1
requests==2.21.0